Changed to micro stepping scheme

+++ lib/Stepper/Stepper.h
@@ -4,7 +4,7 @@
 class Stepper
 {
 public:
-  Stepper(int16_t pin_1, int16_t pin_2, int16_t pin_3, int16_t pin_4);
+  Stepper(uint8_t pin_1, uint8_t pin_2, uint8_t pin_3, uint8_t pin_4);
   void handle();
   void move_abs(int16_t pos);
   void move_rel(int16_t steps);
@@ -19,10 +19,10 @@ private:
   int16_t step_act;
   int16_t step_target;
   // motor pin numbers:
-  int16_t motor_pin_1;
-  int16_t motor_pin_2;
-  int16_t motor_pin_3;
-  int16_t motor_pin_4;
+  uint8_t motor_pin_1;
+  uint8_t motor_pin_2;
+  uint8_t motor_pin_3;
+  uint8_t motor_pin_4;
 
   unsigned long step_delay;
   unsigned long step_next;
